requirements.txt
# Academicon Code Assistant - OpenRouter Edition
# Multi-Agent RAG System with GPU-Accelerated Embeddings
# Python 3.11+

# ============================================================================
# CORE LLAMA INDEX ECOSYSTEM
# ============================================================================
# Vector database and retrieval
llama-index-core>=0.15.0
llama-index-embeddings-huggingface>=0.2.0

# LLM Integrations
llama-index-llms-openai-like>=0.2.0
llama-index-vector-stores-chroma>=0.1.0

# ============================================================================
# VECTOR DATABASE & SEARCH
# ============================================================================
chromadb>=0.5.0,<0.6.0              # Persistent vector database
rank-bm25>=0.2.2                    # BM25 for hybrid search

# ============================================================================
# EMBEDDINGS & TRANSFORMERS
# ============================================================================
sentence-transformers>=3.0.0         # Nomic Embed model
transformers>=4.40.0                # Hugging Face transformers
huggingface-hub>=0.21.0             # Model downloading

# ============================================================================
# PYTORCH & GPU SUPPORT
# ============================================================================
torch>=2.1.0                        # PyTorch with CUDA support
torchvision>=0.16.0                 # Vision utilities
torchaudio>=2.1.0                   # Audio utilities

# ============================================================================
# WEB UI & FRONTEND
# ============================================================================
gradio>=4.30.0                      # Web UI framework
aiofiles>=23.0.0                    # Async file operations

# ============================================================================
# CONFIGURATION & ENVIRONMENT
# ============================================================================
pydantic>=2.5.0                     # Data validation
pydantic-settings>=2.1.0            # Settings management
python-dotenv>=1.0.0                # .env file support

# ============================================================================
# UTILITIES & HELPERS
# ============================================================================
tenacity>=8.2.0                     # Retry logic for API calls
ast-comments>=1.0.0                 # Code parsing

# ============================================================================
# TESTING & DEVELOPMENT
# ============================================================================
pytest>=7.4.0                       # Testing framework
pytest-asyncio>=0.22.0              # Async test support
pytest-cov>=4.1.0                   # Coverage reporting
pytest-mock>=3.12.0                 # Mocking utilities

# ============================================================================
# OPTIONAL: GPU ACCELERATION (ONNX Runtime)
# ============================================================================
# onnxruntime-gpu>=1.19.0           # Uncomment for ONNX optimization
